backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.11
python-dotenv==1.0.0
pymongo==4.4.0
numpy==1.26.2
scikit-learn==1.3.2
opencv-python==4.8.0.76
insightface==0.8.3
face_recognition==1.3.1
python-multipart==0.0.6
